Fix signed vs unsigned: rpmstrpool
Use unsigned int for all kind of counts

Related: #3226

+++ rpmio/rpmstrpool.cc
@@ -26,9 +26,9 @@ struct poolHashBucket_s {
 };
 
 struct poolHash_s {
-    int numBuckets;
+    size_t numBuckets;
     poolHashBucket * buckets;
-    int keyCount;
+    size_t keyCount;
 };
 
 struct rpmstrPool_s {
@@ -130,7 +130,7 @@ static void poolHashResize(rpmstrPool pool, int numBuckets)
     poolHash ht = pool->hash;
     poolHashBucket * buckets = (poolHashBucket *)xcalloc(numBuckets, sizeof(*ht->buckets));
 
-    for (int i=0; i<ht->numBuckets; i++) {
+    for (size_t i=0; i<ht->numBuckets; i++) {
         if (!ht->buckets[i].keyid) continue;
         unsigned int keyHash = rstrhash(id2str(pool, ht->buckets[i].keyid));
         for (unsigned int j=0;;j++) {
@@ -198,9 +198,9 @@ static poolHash poolHashFree(poolHash ht)
 static void poolHashPrintStats(rpmstrPool pool)
 {
     poolHash ht = pool->hash;
-    int i;
-    int collisions = 0;
-    int maxcollisions = 0;
+    size_t i;
+    unsigned int collisions = 0;
+    unsigned int maxcollisions = 0;
 
     for (i=0; i<ht->numBuckets; i++) {
         unsigned int keyHash = rstrhash(id2str(pool, ht->buckets[i].keyid));
@@ -213,8 +213,8 @@ static void poolHashPrintStats(rpmstrPool pool)
             }
         }
     }
-    fprintf(stderr, "Hashsize: %i\n", ht->numBuckets);
-    fprintf(stderr, "Keys: %i\n", ht->keyCount);
+    fprintf(stderr, "Hashsize: %li\n", ht->numBuckets);
+    fprintf(stderr, "Keys: %li\n", ht->keyCount);
     fprintf(stderr, "Collisions: %i\n", collisions);
     fprintf(stderr, "Max collisions: %i\n", maxcollisions);
 }
@@ -232,7 +232,7 @@ static void rpmstrPoolRehash(rpmstrPool pool)
 	pool->hash = poolHashFree(pool->hash);
 
     pool->hash = poolHashCreate(sizehint);
-    for (int i = 1; i <= pool->offs_size; i++)
+    for (size_t i = 1; i <= pool->offs_size; i++)
 	poolHashAddEntry(pool, id2str(pool, i), i);
 }
 
@@ -264,7 +264,7 @@ rpmstrPool rpmstrPoolFree(rpmstrPool pool)
 	poolHashPrintStats(pool);
     poolHashFree(pool->hash);
     free(pool->offs);
-    for (int i=1; i<=pool->chunks_size; i++) {
+    for (size_t i=1; i<=pool->chunks_size; i++) {
 	pool->chunks[i] = _free(pool->chunks[i]);
     }
     free(pool->chunks);
